+++ src/components/registerForm/cardNumber/CardNumber.tsx
@@ -1,14 +1,8 @@
-import React, {
-  ForwardedRef,
-  forwardRef,
-  useImperativeHandle,
-  useRef,
-  useState,
-} from "react";
-import Input from "./common/Input";
+import React, { forwardRef, useState } from "react";
+import Input from "src/components/common/Input";
 import styled, { css } from "styled-components";
-import FormLabel from "./common/FormLabel";
-import ErrorSpan from "./common/ErrorSpan";
+import FormLabel from "src/components/common/FormLabel";
+import ErrorSpan from "src/components/common/ErrorSpan";
 import { ONLY_NUMBER_REGEXP } from "src/utils/regexp";
 
 export interface CardNumberObj {
