+++ src/components/registerForm/cardRegisterForm/CardRegisterForm.stories.tsx
@@ -1,0 +1,16 @@
+import type { Meta, StoryObj } from "@storybook/react";
+import React from "react";
+import CardRegisterForm from "./CardRegisterForm";
+
+const cardRegister = {
+  component: CardRegisterForm,
+  title: "CardRegisterForm Input",
+} satisfies Meta<typeof CardRegisterForm>;
+
+export default cardRegister;
+
+type Story = StoryObj<typeof cardRegister>;
+
+export const Example = {
+  args: {},
+} satisfies Story;
