+++ src/components/registerForm/securityCode/SecurityCode.stories.tsx
