+++ src/components/registerForm/securityCode/SecurityCode.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
-import FormLabel from "./common/FormLabel";
-import Input from "./common/Input";
+import FormLabel from "src/components/common/FormLabel";
+import Input from "src/components/common/Input";
 import { ONLY_NUMBER_REGEXP } from "src/utils/regexp";
 import styled, { css } from "styled-components";
